Some attempts to try to fix frame timer, because it affect movement and stuff

+++ FrameTimer.cpp
@@ -8,7 +8,7 @@ void FrameTimer::init(int fps)
 
     requestedFps = fps;
 
-    intervalsPerFrame = (float)timerFreq.QuadPart / (float)requestedFps;
+    intervalsPerFrame = (float)(timerFreq.QuadPart / requestedFps);
 }
 
 int FrameTimer::framesToUpdate()
